refactor: `debug` into separate module
